SVN Changes: Error handeling in StockSummary. DataNotAvailable exception added. Finance.Web.Example commandline application on how one might use the Finance.Web library.
Notes:
A semi stable class of StockSummary has been finished. Some more exception handling might need to be done. At this point
it is usable. Please keep in mind that the formats not be what your expecting so please play around with it and respond
back to me on what your thoughts are and improvements you think should be made. 
StockSummary object pretty much returns all the important information needed from a security on google's finance page as
a returned object. The DataNotAvailable exception was added for times when the data is not available, such as when the 
EPS is '-'. 

Additional Work Needed:
Integration objects between Finance and Finance.Web libraries. 
Need to implement proxy support.
Need to finish implementing the HistoricalPrices, RelatedCompanies, StockFinancials, StockNews objects